% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coglasso.R
\name{gen_hpars}
\alias{gen_hpars}
\title{Generate combinations of hyperparameters for \code{coglasso()}}
\usage{
gen_hpars(
  S = NULL,
  p = NULL,
  lambda_w = NULL,
  lambda_b = NULL,
  c = NULL,
  nlambda_w = NULL,
  nlambda_b = NULL,
  nc = NULL,
  c_max = NULL,
  lambda_w_min_ratio = NULL,
  lambda_b_min_ratio = NULL,
  c_min_ratio = NULL,
  getc = NULL
)
}
\arguments{
\item{S}{The empirical Pearson's correlation matrix of the multi-omics data set.}

\item{p}{The total number of variables in the multi-omics data set.}

\item{lambda_w}{A vector of values for the parameter \eqn{λ_w}, the
penalization parameter for the "within" interactions. Overrides
\code{nlambda_w}.}

\item{lambda_b}{A vector of values for the parameter \eqn{λ_b}, the
penalization parameter for the "between" interactions. Overrides
\code{nlambda_b}.}

\item{c}{A vector of values for the parameter \eqn{c}, the weight given to
collaboration. Overrides \code{nc}.}

\item{nlambda_w}{The number of requested \eqn{λ_w} parameters to explore. A
sequence of size \code{nlambda_w} of \eqn{λ_w} parameters will be generated.
Defaults to 8. Ignored when \code{lambda_w} is set by the user.}

\item{nlambda_b}{The number of requested \eqn{λ_b} parameters to explore. A
sequence of size \code{nlambda_b} of \eqn{λ_b} parameters will be generated.
Defaults to 8. Ignored when \code{lambda_b} is set by the user.}

\item{nc}{The number of requested \eqn{c} parameters to explore. A sequence
of size \code{nc} of \eqn{c} parameters will be generated. Defaults to 8.
Ignored when \code{c} is set by the user.}

\item{c_max}{The greatest generated \eqn{c}. Defaults to 10. Ignored when \code{c}
is set by the user.}

\item{lambda_w_min_ratio}{The ratio of the smallest generated \eqn{λ_w} over
the greatest generated \eqn{λ_w}. Defaults to 0.1. Ignored when \code{lambda_w}
is set by the user.}

\item{lambda_b_min_ratio}{The ratio of the smallest generated \eqn{λ_b} over
the greatest generated \eqn{λ_b}. Defaults to 0.1. Ignored when \code{lambda_b}
is set by the user.}

\item{c_min_ratio}{The ratio of the smallest generated \eqn{c} over the
greatest generated \eqn{c}. Defaults to 0.1. Ignored when \code{c} is set by the
user.}

\item{getc}{Parameter for testing purposes. If TRUE, the returned table contains a column with \emph{c} values rather than \emph{α} values.}
}
\value{
\code{gen_hpars()} returns a list of four elements:
\itemize{
\item A table of all the combinations of hyperparameters given as input to \code{coglasso()}, with \eqn{α(λ_w+λ_b)} being the key to sort rows.
\item A numerical vector with all the generated \eqn{λ_w}.
\item A numerical vector with all the generated \eqn{λ_b}.
\item A numerical vector with all the generated \eqn{c}.
}
}
\description{
\code{gen_hpars()} generates an ordered table of all the combinations of the hyperparameters given as input to \code{coglasso()}
}
\examples{
# Typical usage: set the number of hyperparameters to explore
cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3)

}
