% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStARS.R
\name{xestars}
\alias{xestars}
\title{Efficient stability selection of the best \code{coglasso} network}
\usage{
xestars(
  coglasso_obj,
  stars_thresh = 0.1,
  stars_subsample_ratio = NULL,
  rep_num = 20,
  max_iter = 10,
  old_sampling = FALSE,
  light = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{coglasso_obj}{The object of \code{S3} class \code{coglasso} returned by \code{coglasso()}.}

\item{stars_thresh}{The threshold set for variability of the explored
networks at each iteration of the algorithm. The \eqn{\lambda_w} or the
\eqn{\lambda_b} associated to the most stable network before the threshold
is overcome is selected.}

\item{stars_subsample_ratio}{The proportion of samples in the multi-omics
data set to be randomly subsampled to estimate the variability of the
network under the given hyperparameters setting. Defaults to 80\% when the
number of samples is smaller than 144, otherwise it defaults to
\eqn{\frac{10}{n}\sqrt{n}}.}

\item{rep_num}{The amount of subsamples of the multi-omics data set used to
estimate the variability of the network under the given hyperparameters
setting. Defaults to 20.}

\item{max_iter}{The greatest number of times the algorithm is allowed to
choose a new best \eqn{\lambda_w}. Defaults to 10.}

\item{old_sampling}{Perform the same subsampling \code{xstars()} would if set to
TRUE. Makes a difference with bigger data sets, where computing
a correlation matrix could take significantly longer. Defaults to FALSE.}

\item{light}{Do not store the "merged" matrixes recording average variability
of each edge, making the algorithm more memory efficient, if set to TRUE.
Defaults to TRUE.}

\item{verbose}{Print information regarding the progress of the selection
procedure on the console.}
}
\value{
\code{xestars()} returns an object of \code{S3} class \code{select_coglasso}
containing the results of the selection
procedure, built upon the object of \code{S3} class \code{coglasso} returned by \code{coglasso()}.
\itemize{
\item ... are the same elements returned by \code{\link[=coglasso]{coglasso()}}.
\item \code{opt_adj} is a list of the adjacency matrices finally selected for each
\eqn{c} parameter explored.
\item \code{opt_variability} is a numerical vector containing the variabilities
associated to the adjacency matrices in \code{opt_adj}.
\item \code{opt_index_lw} and \code{opt_index_lb} are integer vectors containing the
index of the selected \eqn{\lambda_w}s (or \eqn{\lambda_b}s) for each
\eqn{c} parameters explored.
\item \code{opt_lambda_w} and \code{opt_lambda_b} are vectors containing the selected
\eqn{\lambda_w}s (or \eqn{\lambda_b}s) for each \eqn{c} parameters
explored.
\item \code{sel_index_c}, \code{sel_index_lw} and \code{sel_index_lb} are the indexes of the
final selected parameters \eqn{c}, \eqn{\lambda_w} and \eqn{\lambda_b}
leading to the most stable sparse network.
\item \code{sel_c}, \code{sel_lambda_w} and \code{sel_lambda_b} are the final selected
parameters \eqn{c}, \eqn{\lambda_w} and \eqn{\lambda_b} leading to the most
stable sparse network.
\item \code{sel_adj} is the adjacency matrix of the final selected network.
\item \code{sel_density} is the density of the final selected network.
\item \code{sel_icov} is the inverse covariance matrix of the final selected network.
\item \code{call} is the matched call.
\item \code{method} is the chosen model selection method. Here, it is "xestars".
\item \code{merge_lw} and \code{merge_lb} are returned only if \code{light} is set to FALSE.
They are lists with as many elements as the number of
\eqn{c} parameters explored. Every element is a "merged" adjacency matrix,
the average of all the adjacency matrices estimated  for those specific
\eqn{c} and the selected \eqn{\lambda_w} (or \eqn{\lambda_b}) values
across all the subsampling in the last path explored before convergence,
the one when the final combination of \eqn{\lambda_w} and \eqn{\lambda_b}
is selected for the given \eqn{c} value.
}
}
\description{
\code{xestars()} provides a more efficient and lighter implementation
than \code{xstars()} to select the combination of hyperparameters given to
\code{coglasso()} yielding the most stable, yet sparse network. Stability is
computed upon network estimation from multiple subsamples of the multi-omics
data set, allowing repetition. Subsamples are collected for a fixed amount of
times (\code{rep_num}), and with a fixed proportion of the total number of samples
(\code{stars_subsample_ratio}).
}
\details{
\emph{eXtended Efficient StARS} (\emph{XEStARS}) is a more efficient and memory-light version of
\emph{XStARS}, the adaptation for \emph{collaborative graphical regression} of the method
published by Liu, H. \emph{et al.} (2010): Stability Approach to Regularization
Selection (StARS). \emph{StARS} was developed for network estimation regulated by
a single penalty parameter, while \emph{collaborative graphical lasso} needs to
explore three different hyperparameters. In particular, two of these are
penalty parameters with a direct influence on network sparsity, hence on
stability. For every \eqn{c} parameter, \code{xestars()} explores one of the two
penalty parameters (\eqn{\lambda_w} or \eqn{\lambda_b}), keeping the other
one fixed at its previous best estimate, using the normal, one-dimentional
\emph{StARS} approach, until finding the best couple. What makes it more efficient
than \code{xstars()} is that the stability check that in the original algorithm
(even in the original \emph{StARS}) is performed for every \eqn{\lambda_w} or
\eqn{\lambda_b} value, is implemented here as a \emph{stopping criterion}. This
reduces sensibly the number of iterations before convergence. It then selects
the \eqn{c} parameter for which the best (\eqn{\lambda_w}, \eqn{\lambda_b})
couple yielded the most stable, yet sparse network. \cr
The original \emph{XStARS} computes a new subsampling for every time the algorithm
switches from optimizing the two\eqn{\lambda_w} and \eqn{\lambda_b}, and for
every \eqn{c}. This does not allow to compare the hyperparameters on an equal
ground, and can slow the selection down with bigger data set or a larger
hyperparameter space. To allow a fairer (and faster) comparison among
different optimizations, the \code{old_sampling} parameter has been implemented.
If set to TRUE, the subsampling is the same one \code{xstars()} would perform.
Otherwise the subsampling is performed at the beginning of the algorithm once
and for all its iterations. \cr
To allow \code{xestars()} to be more memory light, the \code{light} parameter has been
implemented. If set to TRUE and the "merged" matrixes traditionally returned
by both \emph{StARS} and \emph{XStARS} are not returned.
}
\examples{
cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
\donttest{
# Takes less than five seconds
sel_cg <- xestars(cg, verbose = FALSE)
}
}
