% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coglasso-package.R
\docType{package}
\name{coglasso-package}
\alias{coglasso-package}
\alias{_PACKAGE}
\title{coglasso: Collaborative Graphical Lasso - Multi-Omics Network Reconstruction}
\description{
An algorithm for network reconstruction from multi-omics data sets joining the principles of the graphical lasso by Friedman et al. (2008) and collaborative regression by Gross and Tibshirani (2015).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrQuestion/coglasso}
  \item \url{https://drquestion.github.io/coglasso/}
  \item Report bugs at \url{https://github.com/DrQuestion/coglasso/issues}
}

}
\author{
\strong{Maintainer}: Alessio Albanese \email{alessio.albanese@wur.nl} (\href{https://orcid.org/0000-0003-1783-5613}{ORCID})

}
\keyword{internal}
