% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coglasso-package.R
\docType{package}
\name{coglasso-package}
\alias{coglasso-package}
\alias{_PACKAGE}
\title{coglasso: Collaborative Graphical Lasso - Multi-Omics Network Reconstruction}
\description{
Reconstruct networks from multi-omics data sets with the collaborative graphical lasso (coglasso) algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrQuestion/coglasso}
  \item \url{https://drquestion.github.io/coglasso/}
  \item Report bugs at \url{https://github.com/DrQuestion/coglasso/issues}
}

}
\author{
\strong{Maintainer}: Alessio Albanese \email{alessio.albanese@wur.nl} (\href{https://orcid.org/0000-0003-1783-5613}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Pariya Behrouzi \email{pariya.behrouzi@wur.nl} (\href{https://orcid.org/0000-0001-6762-5433}{ORCID})
}

}
\keyword{internal}
