% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_omics_SD}
\alias{multi_omics_SD}
\title{Multi-omics dataset of sleep deprivation in mouse}
\format{
## `multi_omics_SD`
A data frame with 30 rows and 238 variables (162 transcripts and 76 metabolites):
\describe{
  \item{Plin4 to Tfrc}{log2 CPM values of 162 transcripts in mouse cortex
   under sleep deprivation (-4.52--10.46)}
  \item{Ala to SM C24:1}{abundance values of 76 metabolites (0.02--1112.67)}
}
}
\source{
<https://doi.org/10.1038/s41597-019-0171-x>

<https://figshare.com/articles/dataset/Input_data_for_systems_genetics_of_sleep_regulation/7797434>
}
\usage{
multi_omics_SD
}
\description{
A dataset containing transcript and metabolite values analysed in Albanese et
 al. 2023, subset of the multi-omics data set published in  Jan, M., Gobet, 
 N., Diessler, S. et al. A multi-omics digital research object for the 
 genetics of sleep regulation. Sci Data 6, 258 (2019).
 
`multi_omics_SD_small` is a smaller version, limited to Cirbp and the 
 transcripts and metabolites belonging to the Cirbp's neighborhood described 
 in Albanese et al. 2023
}
\keyword{datasets}
