% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XStARS.R
\name{stars_coglasso}
\alias{stars_coglasso}
\title{Stability selection of the best \code{coglasso} network}
\usage{
stars_coglasso(
  coglasso_obj,
  stars_thresh = 0.1,
  stars_subsample_ratio = NULL,
  rep_num = 20,
  max_iter = 10,
  verbose = FALSE
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{stars_coglasso()} was deprecated in favor of \code{\link[=xstars]{xstars()}} as the function name
does not reflect properly the method it implements. The new function will
also allow new memory-lighter and faster possibilities.
}
\examples{
cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
\donttest{
# Deprecated, use xstars() instead. Takes around one minute
sel_cg <- stars_coglasso(cg, verbose = FALSE)
# ->
sel_cg <- xstars(cg, verbose = FALSE)
}
}
\keyword{internal}
